Added Gui, switched to ArrayList and re-structured the Code
After seeking advice from Sabine as to the architecture of our code and
the way we're headed with it, she suggested a quick restructure in
order to progress better and gave me a good head start. I restructured
the code in order to implement new code easily in the next steps and be
able to grow the code comfortably until the final. The new structure
should allow us more flexibility in iterations and better control for
adjustments and finding our variables. I also added the Gui. We now
have a couple functional knobs to start testing out with before we see
what exactly we want the user to control and manipulate. We also now
have an Array list rather then an array in order to remove cars. I
commented most things in the sketch itself. I also added a spacebar
toggle to enable and disable debug lines. In the next two days I will
be working on the Disturbance agent and the car design.